requirements.txt
langchain-groq>=0.1.3
fastembed>=0.2.7
langchain==0.3.27
langchain-community>=0.3.31
langchain-openai>=0.3.35
langchain-qdrant>=0.2.1
langfuse>=3.6.2
langgraph>=0.6.10
lark>=1.3.0
llama-index>=0.13.6
llama-index-llms-openai>=0.5.4
markitdown>=0.1.3
openai>=1.106.1
python-dotenv>=1.1.1
qdrant-client>=1.15.1
streamlit>=1.50.0
tiktoken>=0.11.0
unstructured[local-inference]>=0.14.0
# Novos formatos
docx2txt>=0.8
beautifulsoup4>=4.12.0
# Dados estruturados
pandas>=2.0.0
openpyxl>=3.1.0
# Áudio e Vídeo
openai-whisper>=20231117
moviepy>=1.0.3
# Testes
pytest>=8.0.0
pytest-mock>=3.12.0
